# fix: increase wait seconds before reboot