@airkei 2025-06-13
In FMS, when the progress is 100%, it surely notifies the progress to backend immediately.
So this issue might be from another reason, close at the moment.